requirements.txt
Pillow >= 8.3.2
PyYAML >= 5.4.1
requests-oauthlib >= 1.3.0